+++ frameworks/core/components_ng/pattern/linear_split/linear_split_layout_algorithm.cpp
@@ -48,7 +48,7 @@ void LinearSplitLayoutAlgorithm::Measure(LayoutWrapper* layoutWrapper)
         }
     } while (false);
 
-    if (!childrenDragPos_.empty() && childrenDragPos_.size() != layoutWrapper->GetTotalChildCount() + 1) {
+    if (!childrenDragPos_.empty() && childrenDragPos_.size() != layoutWrapper->GetTotalChildCount() + 1U) {
         childrenDragPos_.clear();
     }
 
@@ -254,7 +254,7 @@ void LinearSplitLayoutAlgorithm::LayoutColumnSplit(LayoutWrapper* layoutWrapper,
     auto padding = layoutWrapper->GetLayoutProperty()->CreatePaddingAndBorder();
     auto parentWidth = layoutWrapper->GetGeometryNode()->GetFrameSize().Width() - padding.Width();
     bool isFirstSetPos = false;
-    if (!childrenDragPos_.empty() && childrenDragPos_.size() != layoutWrapper->GetTotalChildCount() + 1) {
+    if (!childrenDragPos_.empty() && childrenDragPos_.size() != layoutWrapper->GetTotalChildCount() + 1U) {
         childrenDragPos_.clear();
     }
     if (childrenDragPos_.empty()) {
@@ -304,7 +304,7 @@ void LinearSplitLayoutAlgorithm::LayoutColumnSplit(LayoutWrapper* layoutWrapper,
 void LinearSplitLayoutAlgorithm::ColumnSplitChildConstrain(LayoutWrapper* layoutWrapper,
     const RefPtr<LayoutWrapper>& item, int32_t index)
 {
-    if (index >= childrenConstrains_.size()) {
+    if (static_cast<uint32_t>(index) >= childrenConstrains_.size()) {
         return;
     }
     const auto [startMargin, endMargin] = GetDividerMargin(layoutWrapper);
